requirements.txt
Flask==3.0.3
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
